WIP
- changes to struct & traits
